dblib/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
alembic
python-dotenv
